#22: Add PaginationListener
